@@ -5,17 +5,17 @@ class Question extends StatelessWidget {
   final String questionText; 
  
 //constructor 
-  Question(this.questionText); 
+  const Question(this.questionText, {super.key}); 
  
 //Build method 
   @override 
   Widget build(BuildContext context) { 
     return Container( 
       width: double.infinity, 
-      margin: EdgeInsets.all(10), 
+      margin: const EdgeInsets.all(10), 
       child: Text( 
         questionText, 
-        style: TextStyle(fontSize: 20, color: Colors.red), 
+        style: const TextStyle(fontSize: 20, color: Colors.red), 
         textAlign: TextAlign.center, 
       ), 
     ); 
